+++ tests/mortar/manager_05.cc
@@ -17,7 +17,7 @@ main()
   const std::vector<unsigned int> n_subdivisions      = {10, 3};
   const unsigned int              n_quadrature_points = 3;
   const std::vector<double>       radius              = {1.2, 0.4};
-  const double height_min = radius[1] / n_subdivisions[1] * 0.5;
+  const std::vector<double> stage_heights = {0.0, 0.4 / 3.0, 0.8 / 3.0, 0.4};
 
   // cell angle variation
   const double delta = 2 * numbers::PI / n_subdivisions[0];
@@ -32,7 +32,7 @@ main()
                                                radius,
                                                QGauss<dim>(n_quadrature_points),
                                                rotate,
-                                               height_min);
+                                               stage_heights);
 
       std::cout << manager.get_n_mortars() << std::endl;
       std::cout << manager.get_n_total_mortars() << std::endl;